fixed the schema vs model misnaming
